Create README.md
ProduKkt Kreative Kommunicationz mobile platform "kO-LaBz" is in its infancy. Daarkynspirit / Spiritdancer is the founder and content creator and curator. ProduKkt Kontent Labz (kO-LaBz) is in focus right now and learning all parts of a 1 man mobile content platform with Social Media and online Monetization as the immediate goal.
Imagine a lifelong love affair with art that creates its own abundance, where you wake up anywhere in the world and that is your office, free of geographical limitations and limitless in it's growth potential. 

"Imagine yourself as surrounded by the conditions you intend to produce."
-Wayne Dyer

"In the beginners mind there are many possibilities , in the experts mind there are few."
-Zen Proverb

"The journey of a 1000 miles begins with a single step."
-Lau Tzu